refactor(Actions): Refactor ActionStore
Meh, Aurelia's Store wants to couple the actions into the store (the
dispatch method wants the action name and the payload. I was hoping the
we could actually decouple the whole action into single object (which
would know its name and have a generic method for mutating the state
without needing the store in the middle to do that) that could be simply
passed to store's dispatch method. Maybe there is a reason for this
since Aurelia's store covers much bigger slew of use cases. Maybe in the
future I will implement a bit more simpler store myself, but I will just
hack around the limitiation for now. The Action class implementation is
bit awkward since we have to pass the payload to parent class
constructor and then obtain it again as argument in the method of that
same action class. Well, should hopefully work for now.